Feature/rider

+++ walking-beans-backend/src/main/java/walking_beans/walking_beans_backend/controller/RiderReviewAPIController.java
@@ -17,7 +17,7 @@ public class RiderReviewAPIController {
     private RiderReviewServiceImpl riderReviewService;
 
     @GetMapping("/star")
-    public float getStarRating(@RequestParam("riderId") int riderId) {
+    public float getStarRating(@RequestParam("riderId") long riderId) {
         log.info("=== /api/riderReview/star?riderId={} ===", riderId);
         return riderReviewService.getRiderReviewRatingAverage(riderId);
     }
